TobiChallenge/requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
blis==0.2.4
botbuilder-core==4.0.0a6
botbuilder-schema==4.0.0a6
botframework-connector==4.0.0a6
boto==2.49.0
boto3==1.9.117
botocore==1.12.117
bz2file==0.98
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
cymem==2.0.2
decorator==4.4.0
Django==2.1.7
djangorestframework==3.9.2
docutils==0.14
gensim==3.7.1
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
isodate==0.6.0
it-core-news-sm==2.1.0
jedi==0.13.3
jmespath==0.9.4
jsonschema==2.6.0
msrest==0.4.29
murmurhash==1.0.2
nltk==3.4
numpy==1.16.2
oauthlib==3.0.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
plac==0.9.6
preshed==2.0.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
smart-open==1.8.0
spacy==2.1.0
srsly==0.0.5
thinc==7.0.4
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.1
wasabi==0.1.4
wcwidth==0.1.7
